series cov, count, round methods
