+++ pandas/tests/series/test_analytics.py
@@ -89,217 +89,11 @@ class TestSeriesAnalytics:
             with pytest.raises(ValueError, match=msg):
                 np.compress(cond, s, out=s)
 
-    def test_round(self, datetime_series):
-        datetime_series.index.name = "index_name"
-        result = datetime_series.round(2)
-        expected = Series(
-            np.round(datetime_series.values, 2), index=datetime_series.index, name="ts"
-        )
-        tm.assert_series_equal(result, expected)
-        assert result.name == datetime_series.name
-
-    def test_numpy_round(self):
-        # See gh-12600
-        s = Series([1.53, 1.36, 0.06])
-        out = np.round(s, decimals=0)
-        expected = Series([2.0, 1.0, 0.0])
-        tm.assert_series_equal(out, expected)
-
-        msg = "the 'out' parameter is not supported"
-        with pytest.raises(ValueError, match=msg):
-            np.round(s, decimals=0, out=s)
-
-    def test_numpy_round_nan(self):
-        # See gh-14197
-        s = Series([1.53, np.nan, 0.06])
-        with tm.assert_produces_warning(None):
-            result = s.round()
-        expected = Series([2.0, np.nan, 0.0])
-        tm.assert_series_equal(result, expected)
-
-    def test_built_in_round(self):
-        s = Series([1.123, 2.123, 3.123], index=range(3))
-        result = round(s)
-        expected_rounded0 = Series([1.0, 2.0, 3.0], index=range(3))
-        tm.assert_series_equal(result, expected_rounded0)
-
-        decimals = 2
-        expected_rounded = Series([1.12, 2.12, 3.12], index=range(3))
-        result = round(s, decimals)
-        tm.assert_series_equal(result, expected_rounded)
-
     def test_prod_numpy16_bug(self):
         s = Series([1.0, 1.0, 1.0], index=range(3))
         result = s.prod()
 
         assert not isinstance(result, Series)
-
-    @td.skip_if_no_scipy
-    def test_corr(self, datetime_series):
-        import scipy.stats as stats
-
-        # full overlap
-        tm.assert_almost_equal(datetime_series.corr(datetime_series), 1)
-
-        # partial overlap
-        tm.assert_almost_equal(datetime_series[:15].corr(datetime_series[5:]), 1)
-
-        assert isna(datetime_series[:15].corr(datetime_series[5:], min_periods=12))
-
-        ts1 = datetime_series[:15].reindex(datetime_series.index)
-        ts2 = datetime_series[5:].reindex(datetime_series.index)
-        assert isna(ts1.corr(ts2, min_periods=12))
-
-        # No overlap
-        assert np.isnan(datetime_series[::2].corr(datetime_series[1::2]))
-
-        # all NA
-        cp = datetime_series[:10].copy()
-        cp[:] = np.nan
-        assert isna(cp.corr(cp))
-
-        A = tm.makeTimeSeries()
-        B = tm.makeTimeSeries()
-        result = A.corr(B)
-        expected, _ = stats.pearsonr(A, B)
-        tm.assert_almost_equal(result, expected)
-
-    @td.skip_if_no_scipy
-    def test_corr_rank(self):
-        import scipy.stats as stats
-
-        # kendall and spearman
-        A = tm.makeTimeSeries()
-        B = tm.makeTimeSeries()
-        A[-5:] = A[:5]
-        result = A.corr(B, method="kendall")
-        expected = stats.kendalltau(A, B)[0]
-        tm.assert_almost_equal(result, expected)
-
-        result = A.corr(B, method="spearman")
-        expected = stats.spearmanr(A, B)[0]
-        tm.assert_almost_equal(result, expected)
-
-        # results from R
-        A = Series(
-            [
-                -0.89926396,
-                0.94209606,
-                -1.03289164,
-                -0.95445587,
-                0.76910310,
-                -0.06430576,
-                -2.09704447,
-                0.40660407,
-                -0.89926396,
-                0.94209606,
-            ]
-        )
-        B = Series(
-            [
-                -1.01270225,
-                -0.62210117,
-                -1.56895827,
-                0.59592943,
-                -0.01680292,
-                1.17258718,
-                -1.06009347,
-                -0.10222060,
-                -0.89076239,
-                0.89372375,
-            ]
-        )
-        kexp = 0.4319297
-        sexp = 0.5853767
-        tm.assert_almost_equal(A.corr(B, method="kendall"), kexp)
-        tm.assert_almost_equal(A.corr(B, method="spearman"), sexp)
-
-    def test_corr_invalid_method(self):
-        # GH PR #22298
-        s1 = pd.Series(np.random.randn(10))
-        s2 = pd.Series(np.random.randn(10))
-        msg = "method must be either 'pearson', 'spearman', 'kendall', or a callable, "
-        with pytest.raises(ValueError, match=msg):
-            s1.corr(s2, method="____")
-
-    def test_corr_callable_method(self, datetime_series):
-        # simple correlation example
-        # returns 1 if exact equality, 0 otherwise
-        my_corr = lambda a, b: 1.0 if (a == b).all() else 0.0
-
-        # simple example
-        s1 = Series([1, 2, 3, 4, 5])
-        s2 = Series([5, 4, 3, 2, 1])
-        expected = 0
-        tm.assert_almost_equal(s1.corr(s2, method=my_corr), expected)
-
-        # full overlap
-        tm.assert_almost_equal(
-            datetime_series.corr(datetime_series, method=my_corr), 1.0
-        )
-
-        # partial overlap
-        tm.assert_almost_equal(
-            datetime_series[:15].corr(datetime_series[5:], method=my_corr), 1.0
-        )
-
-        # No overlap
-        assert np.isnan(
-            datetime_series[::2].corr(datetime_series[1::2], method=my_corr)
-        )
-
-        # dataframe example
-        df = pd.DataFrame([s1, s2])
-        expected = pd.DataFrame([{0: 1.0, 1: 0}, {0: 0, 1: 1.0}])
-        tm.assert_almost_equal(df.transpose().corr(method=my_corr), expected)
-
-    def test_cov(self, datetime_series):
-        # full overlap
-        tm.assert_almost_equal(
-            datetime_series.cov(datetime_series), datetime_series.std() ** 2
-        )
-
-        # partial overlap
-        tm.assert_almost_equal(
-            datetime_series[:15].cov(datetime_series[5:]),
-            datetime_series[5:15].std() ** 2,
-        )
-
-        # No overlap
-        assert np.isnan(datetime_series[::2].cov(datetime_series[1::2]))
-
-        # all NA
-        cp = datetime_series[:10].copy()
-        cp[:] = np.nan
-        assert isna(cp.cov(cp))
-
-        # min_periods
-        assert isna(datetime_series[:15].cov(datetime_series[5:], min_periods=12))
-
-        ts1 = datetime_series[:15].reindex(datetime_series.index)
-        ts2 = datetime_series[5:].reindex(datetime_series.index)
-        assert isna(ts1.cov(ts2, min_periods=12))
-
-    def test_count(self, datetime_series):
-        assert datetime_series.count() == len(datetime_series)
-
-        datetime_series[::2] = np.NaN
-
-        assert datetime_series.count() == np.isfinite(datetime_series).sum()
-
-        mi = MultiIndex.from_arrays([list("aabbcc"), [1, 2, 2, np.nan, 1, 2]])
-        ts = Series(np.arange(len(mi)), index=mi)
-
-        left = ts.count(level=1)
-        right = Series([2, 3, 1], index=[1, 2, np.nan])
-        tm.assert_series_equal(left, right)
-
-        ts.iloc[[0, 3, 5]] = np.nan
-        tm.assert_series_equal(ts.count(level=1), right - 1)
-
-        # GH29478
-        with pd.option_context("use_inf_as_na", True):
-            assert pd.Series([pd.Timestamp("1990/1/1")]).count() == 1
 
     def test_dot(self):
         a = Series(np.random.randn(4), index=["p", "q", "r", "s"])
@@ -606,16 +400,6 @@ class TestSeriesAnalytics:
 
 
 class TestCategoricalSeriesAnalytics:
-    def test_count(self):
-
-        s = Series(
-            Categorical(
-                [np.nan, 1, 2, np.nan], categories=[5, 4, 3, 2, 1], ordered=True
-            )
-        )
-        result = s.count()
-        assert result == 2
-
     @pytest.mark.parametrize(
         "dtype",
         ["int_", "uint", "float_", "unicode_", "timedelta64[h]", "datetime64[D]"],
